MAP-1402 BUILD tests and CommonDataIntegrationTest for smaller files
